Advance project - Add Postgres SQL

+++ secrets-cli/index.js
@@ -1,21 +1,28 @@
 #!/usr/bin/env node
 'use strict'
 
+require('dotenv').config()
+
 const minimist = require('minimist')
-const { createDb } = require('./lib/db')
+const promptly = require('promptly')
+const { createDb } = require('./lib')
 const argv = minimist(process.argv.slice(2))
 
+const prompPassword = () => promptly.password('Enter your password:', { replace: '*' })
+
 async function main () {
-  const db = await createDb()
+  const db = await createDb(process.env.DB_TYPE)
   const command = argv._.shift()
 
   switch (command) {
     case 'users:create':
       try {
-        const { user, pass } = argv
+        const { user } = argv
+        const pass = await prompPassword()
         await db.createUser(user, pass)
         console.info(`user: ${user} created`)
       } catch (err) {
+        console.log(err)
         throw new Error('Can not create user')
       }
       break
@@ -34,15 +41,24 @@ async function main () {
     case 'secrets:create':
       try {
         const { user, name, value } = argv
+        const pass = await prompPassword()
+        const isAuth = await db.authenticate(user, pass)
+        if (!isAuth) throw new Error('Invalid user or password')
+
         await db.createSecret(user, name, value)
         console.log(`secret: ${name} created`)
       } catch (err) {
+        console.log(err)
         throw new Error('Can not create secret')
       }
       break
     case 'secrets:list':
       try {
         const { user } = argv
+        const pass = await prompPassword()
+        const isAuth = await db.authenticate(user, pass)
+        if (!isAuth) throw new Error('Invalid user or password')
+
         const secrets = await db.listSecrets(user)
         secrets.forEach(s => {
           console.log(`- ${s.name}`)
@@ -54,6 +70,11 @@ async function main () {
     case 'secrets:get':
       try {
         const { user, name } = argv
+        const pass = await prompPassword()
+
+        const isAuth = await db.authenticate(user, pass)
+        if (!isAuth) throw new Error('Invalid user or password')
+
         const secret = await db.getSecret(user, name)
         if (!secret) return console.log(`secret: ${name} not found`)
         console.log(`- ${secret.name} = ${secret.value}`)
@@ -64,6 +85,11 @@ async function main () {
     case 'secrets:update':
       try {
         const { user, name, value } = argv
+        const pass = await prompPassword()
+
+        const isAuth = await db.authenticate(user, pass)
+        if (!isAuth) throw new Error('Invalid user or password')
+
         await db.updateSecret(user, name, value)
         console.log(`secret: ${name} updated`)
       } catch (err) {
@@ -73,6 +99,11 @@ async function main () {
     case 'secrets:delete':
       try {
         const { user, name } = argv
+        const pass = await prompPassword()
+
+        const isAuth = await db.authenticate(user, pass)
+        if (!isAuth) throw new Error('Invalid user or password')
+
         await db.deleteSecret(user, name)
         console.log(`secret: ${name} deleted`)
       } catch (err) {
